chore/routing_model: extract out scenario tests
